feat: Support AI Developer Edition connections through the Cloud SQL Auth Proxy (#1108)
Cloud SQL AI Developer Edition instances allow connections through the
SqlDataService. With this change and the corresponding changes in
the Auth Proxy, this will will allow users to connect to AIDE instances.

+++ build.sh
@@ -33,12 +33,21 @@ function clean() {
 
 ## generate - Generates all required files, before building
 function generate() {
+  set -x
   get_protoc
+  generated_pb_go=("internal/mdx/metadata_exchange.pb.go"
+    "internal/sqldatagrpc/sql_data_service_grpc.pb.go"
+    "internal/sqldata/sql_data_service.pb.go"
+    )
 
-  pbFile="internal/mdx/metadata_exchange.pb.go"
-  # Delete the old MDX pb file
-  rm "$pbFile"
+  # Delete the old pb files
+  for pbFile in  "${generated_pb_go[@]}" ; do
+    if [[ -f "$pbFile" ]] ; then
+      rm "$pbFile"
+    fi
+  done
 
+  # Generate MDX protos
   PATH="${SCRIPT_DIR}/.tools/protoc/bin:$PATH" "${SCRIPT_DIR}/.tools/protoc/bin/protoc" \
     --proto_path=. \
     --go_out=. \
@@ -46,9 +55,42 @@ function generate() {
     internal/mdx/metadata_exchange.proto \
     --go_opt=paths=source_relative
 
+  # Generate SqlDataService proto messages
+  PATH="${SCRIPT_DIR}/.tools/protoc/bin:$PATH" "${SCRIPT_DIR}/.tools/protoc/bin/protoc" \
+    -I "${SCRIPT_DIR}/internal/google_apis" \
+    --proto_path=. \
+    --go_out=. \
+    --go_opt=paths=source_relative \
+    internal/sqldata/sql_data_service.proto
+
+  # Generate SqlDataService proto grpc stubs
+  PATH="${SCRIPT_DIR}/.tools/protoc/bin:$PATH" "${SCRIPT_DIR}/.tools/protoc/bin/protoc" \
+    -I "${SCRIPT_DIR}/internal/google_apis" \
+    --proto_path=. \
+    --go-grpc_out=.\
+    --go-grpc_opt=paths=source_relative, \
+    internal/sqldata/sql_data_service.proto
+
+
+  # Move the sql_data_service_grpc.pb.go into a separate directory
+  # so that it may be referenced from a different package
+  dest_file="$SCRIPT_DIR/internal/sqldatagrpc/sql_data_service_grpc.pb.go"
+  mkdir -p "$SCRIPT_DIR/internal/sqldatagrpc"
+  mv "$SCRIPT_DIR/internal/sqldata/sql_data_service_grpc.pb.go" "$dest_file"
+  if [[ $(uname) == "Darwin" ]] ; then
+    sed -i '' 's|^package sqldata$|package sqldatagrpc\nimport sqldatapb "cloud.google.com/go/cloudsqlconn/internal/sqldata"|' "$dest_file"
+    sed -i '' 's/StreamSqlDataRequest/sqldatapb.StreamSqlDataRequest/' "$dest_file"
+    sed -i '' 's/StreamSqlDataResponse/sqldatapb.StreamSqlDataResponse/' "$dest_file"
+  else
+    sed -i 's|^package sqldata$|package sqldatagrpc\nimport sqldatapb "cloud.google.com/go/cloudsqlconn/internal/sqldata"|' "$dest_file"
+    sed -i 's/StreamSqlDataRequest/sqldatapb.StreamSqlDataRequest/' "$dest_file"
+    sed -i 's/StreamSqlDataResponse/sqldatapb.StreamSqlDataResponse/' "$dest_file"
+  fi
+
   # Add the copyright header to the generated protobuf file
-  mv "${pbFile}" "${pbFile}.tmp"
-  cat > "${pbFile}" <<EOF
+  for pbFile in  "${generated_pb_go[@]}" ; do
+    mv "${pbFile}" "${pbFile}.tmp"
+    cat > "${pbFile}" <<EOF
 // Copyright 2025 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -64,8 +106,9 @@ function generate() {
 // limitations under the License.
 
 EOF
-  cat "${pbFile}.tmp"  >> "${pbFile}"
-  rm "${pbFile}.tmp"
+    cat "${pbFile}.tmp"  >> "${pbFile}"
+    rm "${pbFile}.tmp"
+  done
 }
 
 # Download the protoc tool if it's not already installed.
@@ -73,6 +116,7 @@ function get_protoc() {
   # Find the latest version of protoc
   protoc_version=$(curl -s "https://api.github.com/repos/protocolbuffers/protobuf/releases/latest" | jq -r '.tag_name' | sed 's/v//')
   proto_go_version=$(curl -s "https://api.github.com/repos/protocolbuffers/protobuf-go/releases/latest" | jq -r '.tag_name' | sed 's/v//')
+  proto_grpc_go_version=$(curl -s "https://api.github.com/repos/grpc/grpc-go/releases" | jq -r '.[].tag_name' | grep cmd/protoc-gen-go-grpc | sed 's|cmd/protoc-gen-go-grpc/v||' | head -n1)
 
   mkdir -p "$SCRIPT_DIR/.tools"
   versioned_cmd="$SCRIPT_DIR/.tools/protoc-$protoc_version"
@@ -89,9 +133,11 @@ function get_protoc() {
   if [[ "$os" == "darwin" && "$arch" == "arm64" ]] ; then
     protoc_url="https://github.com/protocolbuffers/protobuf/releases/download/v${protoc_version}/protoc-${protoc_version}-osx-aarch_64.zip"
     protoc_go_url="https://github.com/protocolbuffers/protobuf-go/releases/download/v${proto_go_version}/protoc-gen-go.v${proto_go_version}.${os}.${protoc_go_arch}.tar.gz"
+    protoc_go_grpc_url="https://github.com/grpc/grpc-go/releases/download/cmd%2Fprotoc-gen-go-grpc%2Fv${proto_grpc_go_version}/protoc-gen-go-grpc.v${proto_grpc_go_version}.${os}.${protoc_go_arch}.tar.gz"
   elif [[ "$os" == "linux" ]] ; then
     protoc_url="https://github.com/protocolbuffers/protobuf/releases/download/v${protoc_version}/protoc-${protoc_version}-${os}-${arch}.zip"
     protoc_go_url="https://github.com/protocolbuffers/protobuf-go/releases/download/v${proto_go_version}/protoc-gen-go.v${proto_go_version}.${os}.${protoc_go_arch}.tar.gz"
+    protoc_go_grpc_url="https://github.com/grpc/grpc-go/releases/download/cmd%2Fprotoc-gen-go-grpc%2Fv${proto_grpc_go_version}/protoc-gen-go-grpc.v${proto_grpc_go_version}.${os}.${protoc_go_arch}.tar.gz"
   else
     echo "Unsupported protoc platform : $os $arch"
     exit 1
@@ -109,6 +155,12 @@ function get_protoc() {
   tar -zxf proto-go.tar.gz -C "$versioned_cmd/bin"
   rm -rf proto-go.tar.gz
 
+  echo "Downloading protoc-go-grpc v$proto_go_version..."
+  curl -v -sSL "$protoc_go_grpc_url" -o proto-go-grpc.tar.gz
+  mkdir -p "$versioned_cmd"
+  tar -zxf proto-go-grpc.tar.gz -C "$versioned_cmd/bin"
+  rm -rf proto-go-grpc.tar.gz
+
   ln -sf "$versioned_cmd" ".tools/protoc"
 }
 
@@ -120,7 +172,13 @@ function build() {
 
 ## test - Runs local unit tests.
 function test() {
-  go test -v -race -cover -short
+  # Install go tools
+  get_golang_tool 'go-junit-report' 'jstemmer/go-junit-report' 'github.com/jstemmer/go-junit-report/v2'
+  mkdir -p test-results
+
+  go test -v -race -cover -short -json \
+    | .tools/go-junit-report -iocopy -parser gojson -out test-results/unit.xml \
+          | jq -j 'select(.Output) | .Output'
 }
 
 ## e2e - Runs end-to-end integration tests.
@@ -135,7 +193,12 @@ function e2e() {
 # e2e_ci - Run end-to-end integration tests in the CI system.
 #   This assumes that the secrets in the env vars are already set.
 function e2e_ci() {
-  go test -v -race -cover ./e2e_mysql_test.go ./e2e_postgres_test.go ./e2e_sqlserver_test.go | tee test_results.txt
+  get_golang_tool 'go-junit-report' 'jstemmer/go-junit-report' 'github.com/jstemmer/go-junit-report/v2'
+  mkdir -p test-results
+
+  go test -v -race -cover ./e2e_mysql_test.go ./e2e_postgres_test.go ./e2e_sqlserver_test.go -json \
+    | .tools/go-junit-report -iocopy -parser gojson -out test-results/e2e.xml \
+    | jq -j 'select(.Output) | .Output '
 }
 
 # Download a tool using `go install`
@@ -143,7 +206,7 @@ function get_golang_tool() {
   name="$1"
   github_repo="$2"
   package="$3"
-
+  set -x
   # Download goimports tool
   version=$(curl -s "https://api.github.com/repos/$github_repo/tags" | jq -r '.[].name' | head -n 1)
   mkdir -p "$SCRIPT_DIR/.tools"
@@ -157,6 +220,7 @@ function get_golang_tool() {
     fi
     ln -s "$versioned_cmd" "$cmd"
   fi
+  set +x
 }
 
 ## fix - Fixes code format.
@@ -245,11 +309,40 @@ function write_e2e_env(){
     val=$(gcloud secrets versions access latest --project "$TEST_PROJECT" --secret="$secret_name")
     echo "export $env_var_name='$val'"
   done
-  echo "export MYSQL_USER_IAM='$(whoami)'"
-  echo "export POSTGRES_USER_IAM='$(whoami)@google.com'"
+  echo "export MYSQL_USER_IAM='$(iam_user_mysql)'"
+  echo "export POSTGRES_USER_IAM='$(iam_user_pg)'"
   } > "$1"
+}
 
+function iam_user_pg() {
+  # Call iam_user_email
+  # Truncate the suffix `.iam.gserviceaccount.com` if it exists. Otherwise return the email
+  local email
+  local pguser
 
+  email="$(iam_user_email)"
+  pguser="${email%%.iam.gserviceaccount.com}"
+  if [[ -n "$pguser" ]] ; then
+    echo "$pguser"
+  else
+    echo "$email"
+  fi
+
+}
+
+function iam_user_mysql() {
+  # Call iam_user_email
+  # Truncate the part after the @
+  local email
+  local pguser
+
+  email=$(iam_user_email)
+  mysqluser="${email%%@*}"
+  echo "$mysqluser"
+}
+
+function iam_user_email() {
+  gcloud auth list --format json | jq -r '.[] | select (.status == "ACTIVE") | .account'
 }
 
 ## help - prints the help details
@@ -278,3 +371,4 @@ fi
 cd "$SCRIPT_DIR"
 
 "$@"
+
